54.Commit:	changed index_type greschd
DONE:
- changed index_type to uint32_t to get rid of problems with the third power
- made a few tests with mt_mkl_rng
- excluded msk_v1_sim from performance_search (still doesn't halt)

TODO:
- fix msk_v1_sim

+++ src/types.hpp
@@ -9,7 +9,7 @@
 
 namespace mc_potts {
     
-    using index_type = uint16_t;
+    using index_type = uint32_t;
     using spin_ret_type = uint8_t;
     using seed_type = uint32_t;
